@@ -6,7 +6,6 @@ import android.graphics.Color;
 import android.graphics.drawable.ColorDrawable;
 import android.net.VpnService;
 import android.os.Bundle;
-import android.util.Log;
 import android.view.View;
 import android.widget.AdapterView;
 import android.widget.Toast;
@@ -39,6 +38,7 @@ import wang.switchy.hin2n.storage.db.base.model.N2NSettingModel;
 import wang.switchy.hin2n.template.BaseTemplate;
 import wang.switchy.hin2n.template.CommonTitleTemplate;
 import wang.switchy.hin2n.tool.N2nTools;
+import wang.switchy.hin2n.tool.ThreadUtils;
 
 
 /**
@@ -127,67 +127,96 @@ public class ListActivity extends BaseActivity {
                             .setConfirmClickListener(new SweetAlertDialog.OnSweetClickListener() {
                                 @Override
                                 public void onClick(SweetAlertDialog sweetAlertDialog) {
-                                    N2NService.INSTANCE.stop();
+                                    N2NService.INSTANCE.stop(new Runnable() {
+                                        @Override
+                                        public void run() {
+                                            Intent vpnPrepareIntent = VpnService.prepare(ListActivity.this);
+                                            if (vpnPrepareIntent != null) {
+                                                startActivityForResult(vpnPrepareIntent, REQUECT_CODE_VPN);
+                                            } else {
+                                                onActivityResult(REQUECT_CODE_VPN, RESULT_OK, null);
+                                            }
+                                        }
+                                    });
 
                                     mTargetSettingPosition = position;
 
-                                    Intent vpnPrepareIntent = VpnService.prepare(ListActivity.this);
-                                    if (vpnPrepareIntent != null) {
-                                        startActivityForResult(vpnPrepareIntent, REQUECT_CODE_VPN);
-                                    } else {
-                                        onActivityResult(REQUECT_CODE_VPN, RESULT_OK, null);
-                                    }
-
                                     if (currentSettingId != -1) {
-                                        N2NSettingModel currentSettingItem = Hin2nApplication.getInstance().getDaoSession().getN2NSettingModelDao().load((long) currentSettingId);
-                                        if (currentSettingItem != null) {
-                                            currentSettingItem.setIsSelcected(false);
-                                            Hin2nApplication.getInstance().getDaoSession().getN2NSettingModelDao().update(currentSettingItem);
-                                        }
+                                        ThreadUtils.cachedThreadExecutor(new Runnable() {
+                                            @Override
+                                            public void run() {
+                                                N2NSettingModel currentSettingItem = Hin2nApplication.getInstance().getDaoSession().getN2NSettingModelDao().load((long) currentSettingId);
+                                                if (currentSettingItem != null) {
+                                                    currentSettingItem.setIsSelcected(false);
+                                                    Hin2nApplication.getInstance().getDaoSession().getN2NSettingModelDao().update(currentSettingItem);
+                                                }
+                                            }
+                                        });
                                     }
 
                                     for (int i = 0; i < mSettingItemEntities.size(); i++) {
                                         mSettingItemEntities.get(i).setSelected(false);
                                     }
-
-                                    N2NSettingModelDao n2NSettingModelDao = Hin2nApplication.getInstance().getDaoSession().getN2NSettingModelDao();
-                                    mN2NSettingModel = n2NSettingModelDao.load(mSettingItemEntities.get(position).getSaveId());
-                                    mN2NSettingModel.setIsSelcected(true);
-
-                                    n2NSettingModelDao.update(mN2NSettingModel);
-
-                                    mHin2nEdit.putLong("current_setting_id", mN2NSettingModel.getId());
-                                    mHin2nEdit.commit();
-                                    mSettingItemEntities.get(position).setSelected(true);
-                                    mSettingItemAdapter.notifyDataSetChanged();
-
+                                    ThreadUtils.cachedThreadExecutor(new Runnable() {
+                                        @Override
+                                        public void run() {
+                                            N2NSettingModelDao n2NSettingModelDao = Hin2nApplication.getInstance().getDaoSession().getN2NSettingModelDao();
+                                            mN2NSettingModel = n2NSettingModelDao.load(mSettingItemEntities.get(position).getSaveId());
+                                            mN2NSettingModel.setIsSelcected(true);
+                                            n2NSettingModelDao.update(mN2NSettingModel);
+                                            mHin2nEdit.putLong("current_setting_id", mN2NSettingModel.getId());
+                                            mHin2nEdit.commit();
+                                            mSettingItemEntities.get(position).setSelected(true);
+                                            ThreadUtils.mainThreadExecutor(new Runnable() {
+                                                @Override
+                                                public void run() {
+                                                    mSettingItemAdapter.notifyDataSetChanged();
+                                                }
+                                            });
+                                        }
+                                    });
                                     sweetAlertDialog.cancel();
                                 }
                             })
                             .show();
                 } else {
                     if (currentSettingId != -1) {
-                        N2NSettingModel currentSettingItem = Hin2nApplication.getInstance().getDaoSession().getN2NSettingModelDao().load((long) currentSettingId);
-                        if (currentSettingItem != null) {
-                            currentSettingItem.setIsSelcected(false);
-                            Hin2nApplication.getInstance().getDaoSession().getN2NSettingModelDao().update(currentSettingItem);
-                        }
+                        ThreadUtils.cachedThreadExecutor(new Runnable() {
+                            @Override
+                            public void run() {
+                                N2NSettingModel currentSettingItem = Hin2nApplication.getInstance().getDaoSession().getN2NSettingModelDao().load((long) currentSettingId);
+                                if (currentSettingItem != null) {
+                                    currentSettingItem.setIsSelcected(false);
+                                    Hin2nApplication.getInstance().getDaoSession().getN2NSettingModelDao().update(currentSettingItem);
+                                }
+                            }
+                        });
                     }
 
                     for (int i = 0; i < mSettingItemEntities.size(); i++) {
                         mSettingItemEntities.get(i).setSelected(false);
                     }
+                    ThreadUtils.cachedThreadExecutor(new Runnable() {
+                        @Override
+                        public void run() {
+                            N2NSettingModelDao n2NSettingModelDao = Hin2nApplication.getInstance().getDaoSession().getN2NSettingModelDao();
+                            mN2NSettingModel = n2NSettingModelDao.load(mSettingItemEntities.get(position).getSaveId());
+                            mN2NSettingModel.setIsSelcected(true);
 
-                    N2NSettingModelDao n2NSettingModelDao = Hin2nApplication.getInstance().getDaoSession().getN2NSettingModelDao();
-                    mN2NSettingModel = n2NSettingModelDao.load(mSettingItemEntities.get(position).getSaveId());
-                    mN2NSettingModel.setIsSelcected(true);
+                            n2NSettingModelDao.update(mN2NSettingModel);
 
-                    n2NSettingModelDao.update(mN2NSettingModel);
+                            mHin2nEdit.putLong("current_setting_id", mN2NSettingModel.getId());
+                            mHin2nEdit.commit();
+                            mSettingItemEntities.get(position).setSelected(true);
+                            ThreadUtils.mainThreadExecutor(new Runnable() {
+                                @Override
+                                public void run() {
+                                    mSettingItemAdapter.notifyDataSetChanged();
+                                }
+                            });
 
-                    mHin2nEdit.putLong("current_setting_id", mN2NSettingModel.getId());
-                    mHin2nEdit.commit();
-                    mSettingItemEntities.get(position).setSelected(true);
-                    mSettingItemAdapter.notifyDataSetChanged();
+                        }
+                    });
                 }
 
             }
@@ -240,10 +269,11 @@ public class ListActivity extends BaseActivity {
                         }
 
                         N2NSettingModel n2NSettingModel = new N2NSettingModel(null, n2NSettingModelCopy.getVersion(), copyName, n2NSettingModelCopy.getIp(), n2NSettingModelCopy.getNetmask(), n2NSettingModelCopy.getCommunity(),
-                                n2NSettingModelCopy.getPassword(), n2NSettingModelCopy.getSuperNode(), n2NSettingModelCopy.getMoreSettings(), n2NSettingModelCopy.getSuperNodeBackup(),
+                                n2NSettingModelCopy.getPassword(), n2NSettingModelCopy.getDevDesc(), n2NSettingModelCopy.getSuperNode(), n2NSettingModelCopy.getMoreSettings(), n2NSettingModelCopy.getSuperNodeBackup(),
                                 n2NSettingModelCopy.getMacAddr(), n2NSettingModelCopy.getMtu(), n2NSettingModelCopy.getLocalIP(), n2NSettingModelCopy.getHolePunchInterval(),
                                 n2NSettingModelCopy.getResoveSupernodeIP(), n2NSettingModelCopy.getLocalPort(), n2NSettingModelCopy.getAllowRouting(), n2NSettingModelCopy.getDropMuticast(),
-                                n2NSettingModelCopy.isUseHttpTunnel(), n2NSettingModelCopy.getTraceLevel(), false);
+                                n2NSettingModelCopy.isUseHttpTunnel(), n2NSettingModelCopy.getTraceLevel(), false, n2NSettingModelCopy.getGatewayIp(), n2NSettingModelCopy.getDnsServer(),
+                                n2NSettingModelCopy.getEncryptionMode());
                         n2NSettingModelDao1.insert(n2NSettingModel);
 
                         //2.ui update
@@ -287,7 +317,7 @@ public class ListActivity extends BaseActivity {
                                         mSettingItemAdapter.notifyDataSetChanged();
 
                                         if (N2NService.INSTANCE != null && currentSettingId == finalSettingItemEntity.getSaveId()) {
-                                            N2NService.INSTANCE.stop();
+                                            N2NService.INSTANCE.stop(null);
                                         }
 
                                         sweetAlertDialog.cancel();
@@ -311,37 +341,45 @@ public class ListActivity extends BaseActivity {
     @Override
     protected void onResume() {
         super.onResume();
+        ThreadUtils.cachedThreadExecutor(new Runnable() {
+            @Override
+            public void run() {
+                N2NSettingModelDao n2NSettingModelDao = Hin2nApplication.getInstance().getDaoSession().getN2NSettingModelDao();
+                List<N2NSettingModel> n2NSettingModels = n2NSettingModelDao.loadAll();
 
-        N2NSettingModelDao n2NSettingModelDao = Hin2nApplication.getInstance().getDaoSession().getN2NSettingModelDao();
-        List<N2NSettingModel> n2NSettingModels = n2NSettingModelDao.loadAll();
+                N2NSettingModel n2NSettingModel;
+                mSettingItemEntities.clear();
+                for (int i = 0; i < n2NSettingModels.size(); i++) {
+                    n2NSettingModel = n2NSettingModels.get(i);
+                    final SettingItemEntity settingItemEntity = new SettingItemEntity(n2NSettingModel.getName(),
+                            n2NSettingModel.getId(), n2NSettingModel.getIsSelcected());
 
-        N2NSettingModel n2NSettingModel;
-        mSettingItemEntities.clear();
-        for (int i = 0; i < n2NSettingModels.size(); i++) {
-            n2NSettingModel = n2NSettingModels.get(i);
-            final SettingItemEntity settingItemEntity = new SettingItemEntity(n2NSettingModel.getName(),
-                    n2NSettingModel.getId(), n2NSettingModel.getIsSelcected());
+                    settingItemEntity.setOnMoreBtnClickListener(new SettingItemEntity.OnMoreBtnClickListener() {
 
-            settingItemEntity.setOnMoreBtnClickListener(new SettingItemEntity.OnMoreBtnClickListener() {
+                        @Override
+                        public void onClick(int positon) {
+                            Intent intent = new Intent(ListActivity.this, SettingDetailsActivity.class);
+                            intent.putExtra("type", SettingDetailsActivity.TYPE_SETTING_MODIFY);
+                            intent.putExtra("saveId", settingItemEntity.getSaveId());
 
-                @Override
-                public void onClick(int positon) {
-                    Intent intent = new Intent(ListActivity.this, SettingDetailsActivity.class);
-                    intent.putExtra("type", SettingDetailsActivity.TYPE_SETTING_MODIFY);
-                    intent.putExtra("saveId", settingItemEntity.getSaveId());
+                            startActivity(intent);
+                        }
+                    });
+                    mSettingItemEntities.add(settingItemEntity);
 
-                    startActivity(intent);
+                    if (n2NSettingModel.getIsSelcected()) {
+                        mHin2nEdit.putLong("current_setting_id", n2NSettingModel.getId());
+                        mHin2nEdit.commit();
+                    }
                 }
-            });
-            mSettingItemEntities.add(settingItemEntity);
-
-            if (n2NSettingModel.getIsSelcected()) {
-                mHin2nEdit.putLong("current_setting_id", n2NSettingModel.getId());
-                mHin2nEdit.commit();
+                ThreadUtils.mainThreadExecutor(new Runnable() {
+                    @Override
+                    public void run() {
+                        mSettingItemAdapter.notifyDataSetChanged();
+                    }
+                });
             }
-        }
-
-        mSettingItemAdapter.notifyDataSetChanged();
+        });
     }
 
     @Override
