@@ -1,17 +1,15 @@
 package wang.switchy.hin2n.activity;
 
 import android.Manifest;
+import android.content.Context;
 import android.content.Intent;
 import android.content.SharedPreferences;
 import android.content.pm.PackageManager;
-import android.net.Uri;
 import android.net.VpnService;
 import android.os.Build;
 import android.os.Bundle;
-import android.support.v4.app.ActionBarDrawerToggle;
-import android.support.v4.app.ActivityCompat;
-import android.support.v4.widget.DrawerLayout;
-import android.util.Log;
+
+import android.text.TextUtils;
 import android.view.View;
 import android.widget.ImageView;
 import android.widget.LinearLayout;
@@ -19,28 +17,25 @@ import android.widget.RelativeLayout;
 import android.widget.TextView;
 import android.widget.Toast;
 
+import androidx.appcompat.app.ActionBarDrawerToggle;
+import androidx.appcompat.widget.Toolbar;
+import androidx.core.widget.NestedScrollView;
+import androidx.drawerlayout.widget.DrawerLayout;
+
 import com.orhanobut.logger.Logger;
 import com.tencent.bugly.beta.Beta;
-import com.umeng.socialize.ShareAction;
-import com.umeng.socialize.UMShareAPI;
-import com.umeng.socialize.UMShareListener;
-import com.umeng.socialize.bean.SHARE_MEDIA;
-import com.umeng.socialize.media.UMImage;
-import com.umeng.socialize.media.UMWeb;
-import com.zhy.m.permission.MPermissions;
-import com.zhy.m.permission.PermissionDenied;
-import com.zhy.m.permission.PermissionGrant;
-import com.zhy.m.permission.ShowRequestPermissionRationale;
-
+import com.yanzhenjie.permission.Action;
+import com.yanzhenjie.permission.AndPermission;
+import com.yanzhenjie.permission.runtime.Permission;
 import org.greenrobot.eventbus.EventBus;
 import org.greenrobot.eventbus.Subscribe;
 import org.greenrobot.eventbus.ThreadMode;
-
-import cn.pedant.SweetAlert.SweetAlertDialog;
+import java.util.List;
 import wang.switchy.hin2n.Hin2nApplication;
 import wang.switchy.hin2n.R;
 import wang.switchy.hin2n.event.ConnectingEvent;
 import wang.switchy.hin2n.event.ErrorEvent;
+import wang.switchy.hin2n.event.LogChangeEvent;
 import wang.switchy.hin2n.event.StartEvent;
 import wang.switchy.hin2n.event.StopEvent;
 import wang.switchy.hin2n.event.SupernodeDisconnectEvent;
@@ -50,18 +45,23 @@ import wang.switchy.hin2n.service.N2NService;
 import wang.switchy.hin2n.storage.db.base.model.N2NSettingModel;
 import wang.switchy.hin2n.template.BaseTemplate;
 import wang.switchy.hin2n.template.CommonTitleTemplate;
+import wang.switchy.hin2n.tool.IOUtils;
 import wang.switchy.hin2n.tool.N2nTools;
+import wang.switchy.hin2n.tool.ShareUtils;
+import wang.switchy.hin2n.tool.ThreadUtils;
 
 public class MainActivity extends BaseActivity {
 
     private N2NSettingModel mCurrentSettingInfo;
     private RelativeLayout mCurrentSettingItem;
     private TextView mCurrentSettingName;
+    private TextView mLogAction;
+    private NestedScrollView mScrollLogAction;
     private ImageView mConnectBtn;
-    private TextView mSupernodeDisconnectNote;
     private DrawerLayout mDrawerLayout;
     private ActionBarDrawerToggle mActionBarDrawerToggle;
     private LinearLayout mLeftMenu;
+    private String logTxtPath;
 
     private static final int REQUECT_CODE_SDCARD = 1;
     private static final int REQUECT_CODE_VPN = 2;
@@ -103,14 +103,14 @@ public class MainActivity extends BaseActivity {
         if (!EventBus.getDefault().isRegistered(this)) {
             EventBus.getDefault().register(this);
         }
-
         mDrawerLayout = (DrawerLayout) findViewById(R.id.drawer_layout);
 
-        mActionBarDrawerToggle = new ActionBarDrawerToggle(this, mDrawerLayout, R.mipmap.ic_launcher, R.string.open, R.string.close) {
+        mActionBarDrawerToggle = new ActionBarDrawerToggle(this, mDrawerLayout, null, R.string.open, R.string.close) {
             //菜单打开
             @Override
             public void onDrawerOpened(View drawerView) {
                 super.onDrawerOpened(drawerView);
+                drawerView.setClickable(true);
             }
 
             // 菜单关闭
@@ -125,6 +125,8 @@ public class MainActivity extends BaseActivity {
         mLeftMenu = (LinearLayout) findViewById(R.id.ll_menu_left);
 
         mConnectBtn = (ImageView) findViewById(R.id.iv_connect_btn);
+        mLogAction = (TextView) findViewById(R.id.tv_log_action);
+        mScrollLogAction = (NestedScrollView) findViewById(R.id.scroll_log_action);
 
         if (N2NService.INSTANCE == null) {
             mConnectBtn.setImageResource(R.mipmap.ic_state_disconnect);
@@ -149,8 +151,13 @@ public class MainActivity extends BaseActivity {
 
                 EdgeStatus.RunningStatus status = N2NService.INSTANCE == null ? EdgeStatus.RunningStatus.DISCONNECT : N2NService.INSTANCE.getCurrentStatus();
                 if (N2NService.INSTANCE != null && status != EdgeStatus.RunningStatus.DISCONNECT && status != EdgeStatus.RunningStatus.FAILED) {
-                    N2NService.INSTANCE.stop();
+                    /* Asynchronous call */
+                    mConnectBtn.setClickable(false);
+                    mConnectBtn.setImageResource(R.mipmap.ic_state_connect_change);
+                    N2NService.INSTANCE.stop(null);
                 } else {
+                    mConnectBtn.setClickable(false);
+                    mConnectBtn.setImageResource(R.mipmap.ic_state_connect_change);
                     Intent vpnPrepareIntent = VpnService.prepare(MainActivity.this);
                     if (vpnPrepareIntent != null) {
                         startActivityForResult(vpnPrepareIntent, REQUECT_CODE_VPN);
@@ -161,7 +168,6 @@ public class MainActivity extends BaseActivity {
             }
         });
 
-        mSupernodeDisconnectNote = (TextView) findViewById(R.id.tv_supernode_disconnect_note);
 
         mCurrentSettingItem = (RelativeLayout) findViewById(R.id.rl_current_setting_item);
         mCurrentSettingItem.setOnClickListener(new View.OnClickListener() {
@@ -176,10 +182,6 @@ public class MainActivity extends BaseActivity {
 
         initLeftMenu();
 
-        // TODO 高版本日志权限
-//        if (!MPermissions.shouldShowRequestPermissionRationale(MainActivity.this, Manifest.permission.WRITE_EXTERNAL_STORAGE, REQUECT_CODE_SDCARD)) {
-//            MPermissions.requestPermissions(MainActivity.this, REQUECT_CODE_SDCARD, Manifest.permission.WRITE_EXTERNAL_STORAGE);
-//        }
     }
 
     private void initLeftMenu() {
@@ -193,32 +195,23 @@ public class MainActivity extends BaseActivity {
                 Logger.d("shareItem onClick~");
 
                 if (Build.VERSION.SDK_INT >= 23) {
-                    String[] permissionList = new String[]{
-//                            Manifest.permission.WRITE_EXTERNAL_STORAGE,
-                            Manifest.permission.ACCESS_FINE_LOCATION,
-//                            Manifest.permission.CALL_PHONE,
-//                            Manifest.permission.READ_LOGS,
-                            Manifest.permission.READ_PHONE_STATE,
-                            Manifest.permission.READ_EXTERNAL_STORAGE,
-//                            Manifest.permission.SET_DEBUG_APP,
-//                            Manifest.permission.SYSTEM_ALERT_WINDOW,
-//                            Manifest.permission.GET_ACCOUNTS,
-//                            Manifest.permission.WRITE_APN_SETTINGS
-                    };
-//                    String[] DeniedPermissions = new String[]{};
-//                    for (int i = 0; i < permissionList.length; i++) {
-//                        if (ContextCompat.checkSelfPermission(MainActivity.this,
-//                                permissionList[i])
-//                                != PackageManager.PERMISSION_GRANTED) {
-//                            DeniedPermissions
-//                        }
-//                    }
-                    ActivityCompat.requestPermissions(MainActivity.this, permissionList, 123);
-
-
+                    AndPermission.with(MainActivity.this)
+                            .runtime()
+                            .permission(Permission.READ_EXTERNAL_STORAGE, Permission.ACCESS_FINE_LOCATION, Permission.READ_PHONE_STATE)
+                            .onGranted(new Action<List<String>>() {
+                                @Override
+                                public void onAction(List<String> data) {
+                                    ShareUtils.doOnClickShareItem(MainActivity.this);
+                                }
+                            })
+                            .onDenied(new Action<List<String>>() {
+                                @Override
+                                public void onAction(List<String> data) {
+                                    Toast.makeText(MainActivity.this, "I NEED PERMISSIONS!", Toast.LENGTH_SHORT).show();
+                                }
+                            }).start();
                 } else {
-                    doOnClickShareItem();
-
+                    ShareUtils.doOnClickShareItem(MainActivity.this);
                 }
 
             }
@@ -229,12 +222,7 @@ public class MainActivity extends BaseActivity {
         contactItem.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                boolean b = joinQQGroup("5QSK63d7uDivxPW2oCpWHyi7FmE4sAzo");
-                if (!b) {
-                    Intent intent = new Intent(MainActivity.this, WebViewActivity.class);
-                    intent.putExtra(WebViewActivity.WEB_VIEW_TYPE, WebViewActivity.TYPE_WEB_VIEW_CONTACT);
-                    startActivity(intent);
-                }
+                ShareUtils.joinQQGroup(MainActivity.this);
             }
         });
 
@@ -268,64 +256,7 @@ public class MainActivity extends BaseActivity {
         });
     }
 
-    private void doOnClickShareItem() {
 
-        UMWeb umWeb = new UMWeb("https://github.com/switch-iot/hin2n/blob/master/README.md");
-        umWeb.setTitle("Hin2n");
-        umWeb.setThumb(new UMImage(this, R.mipmap.ic_launcher));
-        umWeb.setDescription("N2N is a VPN project that supports p2p.");
-
-        new ShareAction(MainActivity.this)
-                .withMedia(umWeb)
-                .setDisplayList(SHARE_MEDIA.WEIXIN, SHARE_MEDIA.WEIXIN_CIRCLE/**,SHARE_MEDIA.SINA*/)
-                .setCallback(new UMShareListener() {
-                    /**
-                     * @descrption 分享开始的回调
-                     * @param platform 平台类型
-                     */
-                    @Override
-                    public void onStart(SHARE_MEDIA platform) {
-
-                    }
-
-                    /**
-                     * @descrption 分享成功的回调
-                     * @param platform 平台类型
-                     */
-                    @Override
-                    public void onResult(SHARE_MEDIA platform) {
-//                        Toast.makeText(MainActivity.this, "成功了", Toast.LENGTH_LONG).show();
-
-                        Log.e("zhangbzshare", "onResult");
-                    }
-
-                    /**
-                     * @descrption 分享失败的回调
-                     * @param platform 平台类型
-                     * @param t 错误原因
-                     */
-                    @Override
-                    public void onError(SHARE_MEDIA platform, Throwable t) {
-//                        Toast.makeText(MainActivity.this, "失败" + t.getMessage(), Toast.LENGTH_LONG).show();
-                        Log.e("zhangbzshare", "onError : " + t.getMessage());
-
-                        Intent intent = new Intent(MainActivity.this, WebViewActivity.class);
-                        intent.putExtra(WebViewActivity.WEB_VIEW_TYPE, WebViewActivity.TYPE_WEB_VIEW_SHARE);
-                        startActivity(intent);
-
-                    }
-
-                    /**
-                     * @descrption 分享取消的回调
-                     * @param platform 平台类型
-                     */
-                    @Override
-                    public void onCancel(SHARE_MEDIA platform) {
-//                        Toast.makeText(MainActivity.this, "取消了", Toast.LENGTH_LONG).show();
-                        Log.e("zhangbzshare", "onCancel");
-                    }
-                }).open();
-    }
 
 
     @Override
@@ -336,8 +267,6 @@ public class MainActivity extends BaseActivity {
     @Override
     protected void onActivityResult(int requestCode, int resultCode, Intent data) {
         super.onActivityResult(requestCode, resultCode, data);
-
-        //UMShareAPI.get(this).onActivityResult(requestCode, resultCode, data);
 
         if (requestCode == REQUECT_CODE_VPN && resultCode == RESULT_OK) {
             Intent intent = new Intent(MainActivity.this, N2NService.class);
@@ -353,8 +282,9 @@ public class MainActivity extends BaseActivity {
     @Override
     protected void onResume() {
         super.onResume();
-
         SharedPreferences n2nSp = getSharedPreferences("Hin2n", MODE_PRIVATE);
+        logTxtPath = n2nSp.getString("current_log_path", "");
+        showLog();
         Long currentSettingId = n2nSp.getLong("current_setting_id", -1);
         if (currentSettingId != -1) {
             mCurrentSettingInfo = Hin2nApplication.getInstance().getDaoSession().getN2NSettingModelDao().load((long) currentSettingId);
@@ -365,7 +295,6 @@ public class MainActivity extends BaseActivity {
             }
 
             mConnectBtn.setVisibility(View.VISIBLE);
-            mSupernodeDisconnectNote.setVisibility(View.GONE);
             if (N2NService.INSTANCE == null) {
                 mConnectBtn.setImageResource(R.mipmap.ic_state_disconnect);
             } else {
@@ -374,7 +303,6 @@ public class MainActivity extends BaseActivity {
                     mConnectBtn.setImageResource(R.mipmap.ic_state_connect);
                 } else if (status == EdgeStatus.RunningStatus.SUPERNODE_DISCONNECT) {
                     mConnectBtn.setImageResource(R.mipmap.ic_state_supernode_diconnect);
-                    mSupernodeDisconnectNote.setVisibility(View.VISIBLE);
                 } else {
                     mConnectBtn.setImageResource(R.mipmap.ic_state_disconnect);
                 }
@@ -385,12 +313,13 @@ public class MainActivity extends BaseActivity {
     @Override
     protected void onPause() {
         super.onPause();
+        SharedPreferences n2nSp = getSharedPreferences("Hin2n", MODE_PRIVATE);
+        n2nSp.edit().putString("current_log_path", logTxtPath).apply();
     }
 
     @Override
     protected void onDestroy() {
         super.onDestroy();
-
         if (EventBus.getDefault().isRegistered(this)) {
             EventBus.getDefault().unregister(this);
         }
@@ -399,23 +328,31 @@ public class MainActivity extends BaseActivity {
 
     @Subscribe(threadMode = ThreadMode.MAIN)
     public void onStartEvent(StartEvent event) {
-        mConnectBtn.setVisibility(View.VISIBLE);
-        mConnectBtn.setImageResource(R.mipmap.ic_state_connect);
-        mSupernodeDisconnectNote.setVisibility(View.GONE);
+        mConnectBtn.postDelayed(new Runnable() {
+            @Override
+            public void run() {
+                mConnectBtn.setImageResource(R.mipmap.ic_state_connect);
+                mConnectBtn.setClickable(true);
+            }
+        }, 200);
+
     }
 
     @Subscribe(threadMode = ThreadMode.MAIN)
     public void onStopEvent(StopEvent event) {
-        mConnectBtn.setVisibility(View.VISIBLE);
-        mConnectBtn.setImageResource(R.mipmap.ic_state_disconnect);
-        mSupernodeDisconnectNote.setVisibility(View.GONE);
+        mConnectBtn.postDelayed(new Runnable() {
+            @Override
+            public void run() {
+                mConnectBtn.setImageResource(R.mipmap.ic_state_disconnect);
+                mConnectBtn.setClickable(true);
+            }
+        }, 200);
     }
 
     @Subscribe(threadMode = ThreadMode.MAIN)
     public void onErrorEvent(ErrorEvent event) {
         mConnectBtn.setVisibility(View.VISIBLE);
         mConnectBtn.setImageResource(R.mipmap.ic_state_disconnect);
-        mSupernodeDisconnectNote.setVisibility(View.GONE);
 
         Toast.makeText(mContext, getString(R.string.toast_connect_failed), Toast.LENGTH_SHORT).show();
     }
@@ -429,89 +366,39 @@ public class MainActivity extends BaseActivity {
     public void onSupernodeDisconnectEvent(SupernodeDisconnectEvent event) {
         mConnectBtn.setVisibility(View.VISIBLE);
         mConnectBtn.setImageResource(R.mipmap.ic_state_supernode_diconnect);
-        mSupernodeDisconnectNote.setVisibility(View.VISIBLE);
-
         Toast.makeText(mContext, getString(R.string.toast_disconnect_and_retry), Toast.LENGTH_SHORT).show();
     }
 
-    /****************
-     * 发起添加群流程。群号：手机版n2n(hin2n)交流群(769731491) 的 key 为： 5QSK63d7uDivxPW2oCpWHyi7FmE4sAzo
-     * 调用 joinQQGroup(5QSK63d7uDivxPW2oCpWHyi7FmE4sAzo) 即可发起手Q客户端申请加群 手机版n2n(hin2n)交流群(769731491)
-     *
-     * @param key 由官网生成的key
-     * @return 返回true表示呼起手Q成功，返回fals表示呼起失败
-     ******************/
-    public boolean joinQQGroup(String key) {
-        Intent intent = new Intent();
-        intent.setData(Uri.parse("mqqopensdkapi://bizAgent/qm/qr?url=http%3A%2F%2Fqm.qq.com%2Fcgi-bin%2Fqm%2Fqr%3Ffrom%3Dapp%26p%3Dandroid%26k%3D" + key));
-        // 此Flag可根据具体产品需要自定义，如设置，则在加群界面按返回，返回手Q主界面，不设置，按返回会返回到呼起产品界面    //intent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK)
-        try {
-            startActivity(intent);
-            return true;
-        } catch (Exception e) {
-            // 未安装手Q或安装的版本不支持
-            return false;
-        }
+
+    @Subscribe(threadMode = ThreadMode.MAIN)
+    public void onLogChangeEvent(final LogChangeEvent event) {
+//        Toast.makeText(this,"LogChageEvent",Toast.LENGTH_SHORT).show();
+        logTxtPath = event.getTxtPath();
+        showLog();
     }
 
-
-    /**
-     * check permission
-     *
-     * @param requestCode
-     * @param permissions
-     * @param grantResults
-     */
-    @Override
-    public void onRequestPermissionsResult(int requestCode, String[] permissions, int[] grantResults) {
-        MPermissions.onRequestPermissionsResult(this, requestCode, permissions, grantResults);
-        super.onRequestPermissionsResult(requestCode, permissions, grantResults);
-
-        if (requestCode == 123) {
-            for (int i = 0; i < grantResults.length; i++) {
-//                Log.e("zhangbzshare", "permission[" + i + "] = " + permissions[i] + ",grantResult[" + i + "] = " + grantResults[i]);
-                if (grantResults[i] != PackageManager.PERMISSION_GRANTED) {
-                    Toast.makeText(mContext, "Permission Denied", Toast.LENGTH_SHORT).show();
-                    return;
+    private void showLog() {
+        ThreadUtils.cachedThreadExecutor(new Runnable() {
+            @Override
+            public void run() {
+                if(!TextUtils.isEmpty(logTxtPath)){
+                    final String logText = IOUtils.readTxtLimit(logTxtPath,1024*2);
+                    ThreadUtils.mainThreadExecutor(new Runnable() {
+                        @Override
+                        public void run() {
+                            mLogAction.setText(logText);
+                            mLogAction.post(new Runnable() {
+                                @Override
+                                public void run() {
+                                    mScrollLogAction.fullScroll(NestedScrollView.FOCUS_DOWN);
+                                }
+                            });
+                        }
+                    });
                 }
             }
-
-            doOnClickShareItem();
-        }
-
+        });
     }
 
 
-    @PermissionGrant(REQUECT_CODE_SDCARD)
-    public void requestSdcardSuccess() {
-        Toast.makeText(this, "GRANT ACCESS SDCARD!", Toast.LENGTH_SHORT).show();
-    }
-
-    @PermissionDenied(REQUECT_CODE_SDCARD)
-    public void requestSdcardFailed() {
-        Toast.makeText(this, "DENY ACCESS SDCARD!", Toast.LENGTH_SHORT).show();
-//        finish();
-    }
-
-    @ShowRequestPermissionRationale(REQUECT_CODE_SDCARD)
-    public void ShowRequestPermissionRationale() {
-        Toast.makeText(this, "ShowRequestPermissionRationale", Toast.LENGTH_SHORT).show();
-        Logger.d("ShowRequestPermissionRationale");
-
-//        mConnectBtn.setImageResource(R.mipmap.ic_state_supernode_diconnect);
-//        mSupernodeDisconnectNote.setVisibility(View.VISIBLE);
-
-        SweetAlertDialog sweetAlertDialog = new SweetAlertDialog(this, SweetAlertDialog.NORMAL_TYPE);
-        sweetAlertDialog
-                .setTitleText("I need permission!")
-                .setConfirmClickListener(new SweetAlertDialog.OnSweetClickListener() {
-                    @Override
-                    public void onClick(SweetAlertDialog sweetAlertDialog) {
-                        sweetAlertDialog.dismiss();
-                        MPermissions.requestPermissions(MainActivity.this, REQUECT_CODE_SDCARD, Manifest.permission.WRITE_EXTERNAL_STORAGE);
-
-                    }
-                })
-                .show();
-    }
 }
